youtube-blog-generator/requirements.txt
Flask==2.0.1
requests==2.26.0
youtube-transcript-api==0.4.4
python-dotenv==0.19.0
markdown==3.3.4

